requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.33.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
apex=0.1=pypi_0
attrs=23.2.0=pypi_0
bert-score=0.3.13=pypi_0
blas=1.0=mkl
blis=0.7.11=pypi_0
brotli-python=1.0.9=py311h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
catalogue=2.0.10=pypi_0
certifi=2024.2.2=py311h06a4308_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.18.1=pypi_0
colorama=0.4.6=pypi_0
confection=0.1.5=pypi_0
contourpy=1.2.1=pypi_0
cuda-cccl=12.4.127=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-driver-dev=12.5.82=0
cuda-driver-dev_linux-64=12.5.82=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-opencl-dev=12.4.127=0
cuda-profiler-api=12.4.127=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cxxfilt=0.3.0=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
datasets=2.20.0=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
docker-pycreds=0.4.0=pypi_0
easse=0.2.4=dev_0
en-core-web-sm=3.7.1=pypi_0
entmax=1.3=pypi_0
et-xmlfile=1.1.0=pypi_0
evaluate=0.4.3=pypi_0
filelock=3.13.1=py311h06a4308_0
fonttools=4.51.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
ftfy=6.2.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.24.2=pypi_0
idna=3.7=pypi_0
iniconfig=2.0.0=pypi_0
intel-openmp=2021.4.0=pypi_0
jinja2=3.1.3=py311h06a4308_0
joblib=1.4.0=pypi_0
jpeg=9e=h5eee18b_1
jsonargparse=3.13.1=pypi_0
kiwisolver=1.4.5=pypi_0
langcodes=3.4.0=pypi_0
language-data=1.2.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lens-metric=0.2.0=pypi_0
lerc=3.0=h295c915_0
levenshtein=0.25.1=pypi_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcufile=1.10.1.7=0
libcufile-dev=1.10.1.7=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libuv=1.44.2=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lightning-utilities=0.11.7=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
locate=1.1.1=pypi_0
lxml=5.2.1=pypi_0
lz4-c=1.9.4=h6a678d5_1
marisa-trie=1.2.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py311h5eee18b_0
matplotlib=3.8.4=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=pypi_0
mkl-service=2.4.0=py311h5eee18b_0
mkl_fft=1.3.1=py311h30b3d60_0
mkl_random=1.2.2=py311hba01205_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpir=3.0.0=h3e5f119_1
mpmath=1.3.0=py311h06a4308_0
msgpack=1.0.8=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
murmurhash=1.0.10=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=py311h06a4308_0
nltk=3.8.1=pypi_0
numpy=1.24.3=py311hc206e33_0
numpy-base=1.24.3=py311hfd5febd_0
openai=1.28.1=pypi_0
openjpeg=2.4.0=h9ca470c_1
openpyxl=3.1.5=pypi_0
openssl=3.0.13=h7f8727e_2
packaging=24.0=pypi_0
pandas=1.5.3=pypi_0
pillow=10.2.0=py311h5eee18b_0
pip=24.2=pypi_0
platformdirs=4.2.2=pypi_0
plotly=5.22.0=pypi_0
pluggy=1.5.0=pypi_0
portalocker=2.8.2=pypi_0
preshed=3.0.9=pypi_0
protobuf=4.25.5=pypi_0
psutil=6.0.0=pypi_0
pyarrow=16.1.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
pyphen=0.15.0=pypi_0
pysocks=1.7.1=py311h06a4308_0
pytest=8.2.2=pypi_0
python=3.11.9=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-levenshtein=0.25.1=pypi_0
pytorch=2.3.0=py3.11_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-extension=0.2=pypi_0
pytorch-lightning=2.0.9=pypi_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py311h5eee18b_0
rake-nltk=1.0.6=pypi_0
rapidfuzz=3.9.0=pypi_0
readline=8.2=h5eee18b_0
regex=2024.4.16=pypi_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
sacrebleu=2.4.2=pypi_0
sacremoses=0.1.1=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.4.2=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=2.9.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py311h06a4308_0
shellingham=1.5.4=pypi_0
simalign=0.4=pypi_0
six=1.16.0=pyhd3eb1b0_1
smart-open=7.0.4=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
spacy=3.7.5=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.41.2=h5eee18b_0
srsly=2.4.8=pypi_0
stanfordnlp=0.2.0=pypi_0
sympy=1.12=py311h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.12.0=pypi_0
tenacity=8.2.3=pypi_0
textstat=0.7.3=pypi_0
thinc=8.2.5=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.19.1=pypi_0
torchaudio=2.3.0=pypi_0
torchmetrics=0.10.3=pypi_0
torchtriton=2.3.0=py311
torchvision=0.18.0=pypi_0
tqdm=4.66.4=pypi_0
transformers=4.40.0=pypi_0
tseval=1.0=pypi_0
typer=0.12.3=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024.1=pypi_0
unbabel-comet=2.2.2=pypi_0
urllib3=2.2.2=pypi_0
wandb=0.17.4=pypi_0
wasabi=1.1.3=pypi_0
wcwidth=0.2.13=pypi_0
weasel=0.4.1=pypi_0
wheel=0.41.2=py311h06a4308_0
win-inet-pton=1.1.0=pypi_0
wordfreq=3.1.1=pypi_0
wrapt=1.16.0=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
yattag=1.15.2=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
